Separate the DSL code and the cfn_cmd code
 - Moved the generic DSL-related code in dsl.rb
 - Added a new 'raw_template' DSL function that takes
    its input as arguments
 - Added a generate_template function that returns the
    template JSON
